+++ wonambi/widgets/utils.py
@@ -8,6 +8,7 @@ from os.path import dirname, join, realpath
 
 from PyQt5.QtCore import QRectF, QSettings, Qt
 from PyQt5.QtGui import (QBrush,
+                         QPen,
                          QColor,
                          QPainterPath,
                          QPainter,
@@ -124,7 +125,10 @@ class RectMarker(QGraphicsRectItem):
     def paint(self, painter, option, widget):
         color = QColor(self.color)
         painter.setBrush(QBrush(color))
-        painter.setPen(Qt.NoPen)
+        p = QPen()
+        p.setWidth(0)
+        p.setColor(color)
+        painter.setPen(p)
         painter.drawRect(self.marker)
         super().paint(painter, option, widget)
 
@@ -265,9 +269,15 @@ class FormFloat(QLineEdit):
     """Subclass QLineEdit for float to have a more consistent API across
     widgets.
 
+    Parameters
+    ----------
+    significant_digits : int
+        number of significant digits
+
     """
-    def __init__(self, default=None, maxw=None):
+    def __init__(self, default=None, maxw=None, significant_digits=3):
         super().__init__('')
+        self.significant_digits = significant_digits
 
         if default is not None:
             self.set_value(default)
@@ -306,9 +316,13 @@ class FormFloat(QLineEdit):
         ----------
         value : float
             value for the line edit
-
         """
-        self.setText(str(value))
+        if value == '' or value is None:
+            text = ''
+        else:
+            text = ('{:.' + str(self.significant_digits) + 'f}').format(value)
+
+        self.setText(text)
 
     def connect(self, funct):
         """Call funct when the text was changed.